+++ CHANGELOG.md
@@ -31,12 +31,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.1.0 - 2022-02-06
+
+### Added
+
+- Automatic reload window with prompt dialog on configuration change
+- No header displayed when both title and tags are not identified
+
+### Changed
+
+- Default configuration value is `true`
+
 ## v0.0.4 - 2022-02-05
 
 ### Changed
 
-- README.md
-- CHANGELOG.md
+- Updated README.md
+- Updated CHANGELOG.md
 
 ## v0.0.3 - 2022-02-05